Оновлено README.md, та повномасштабний рефакторинг структури файлів
