+++ backend/database/Database.php
@@ -1,0 +1,26 @@
+<?php
+class Database {
+    private $host = 'localhost';
+    private $dbname = 'laptop_market';
+    private $username = 'root';
+    private $password = '';
+    private $pdo;
+
+    public function __construct() {
+        try {
+            $this->pdo = new PDO("mysql:host={$this->host};dbname={$this->dbname};charset=utf8mb4",
+                $this->username,
+                $this->password);
+            $this->pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+            $this->pdo->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
+            $this->pdo->setAttribute(PDO::ATTR_EMULATE_PREPARES, false);
+        } catch (PDOException $e) {
+            die('Не вдалося підключитися до бази даних: ' . $e->getMessage());
+        }
+    }
+
+    public function getConnection() {
+        return $this->pdo;
+    }
+}
+?>
